chore: Move project to subfolder
Building docs and trying to follow the cli structure made me
realize that the main files need to be in a subdirectory in order
to call everything from the main directory.

Also added initial docs, testing Sphinx.  It's mostly a bitch,
documentation is terrible.  Insanity for a product that is meant
for documentation. 😂