update: Final update for v1.5.4 release
Learned lots of lessons with the "release".
Make sure to build and install locally many times with many options
before uploading to stop from uploading so many versions and trying
to push tags all the time.

Don't delete branches or move to the main branch before everything
is sure to be working, or you are committing from the main branch
all the time.  Heed my wise words future self!!